logparser_service/requirements.txt
fastapi
uvicorn[standard]
pandas
logparser3
python-multipart 